+++ samples/browser/scripts/fields.js
@@ -25,6 +25,6 @@
   console.log(person.getAge());
 
   // Error will be raised here:
-  //person.setGender("male");
+  // person.setGender("male");
 
 }());