+++ samples/node/object_factory.js
@@ -2,6 +2,23 @@
 
 var class4js = require("../../lib/class4js.js");
 
+var Entity = $class({
+
+  __construct__: function () {
+    this.__id = 0;
+  },
+
+  id: {
+    get: function () {
+      return this.__id;
+    },
+    set: function (value) {
+      this.__id = value;
+    }
+  }
+
+});
+
 var Person = $class({
   
   __construct__: function () {
@@ -27,9 +44,10 @@ var Person = $class({
     }
   }
 
-});
+}, Entity);
 
 var person = new Person({
+  id: 10,
   name: "John Smith",
   age: 30
 });
@@ -51,6 +69,7 @@ var organisation = new Organisation();
 
 organisation.add(person);
 organisation.add({
+  id: 24,
   name: "Peter Joe",
   age: 42
 });