+++ src/class.js
@@ -98,17 +98,36 @@ var Class = Object.create(null, {
    */
   __descriptorsAreEqual: { 
     value: function (property, source, target) {
-      var properties = Object.keys(target);
-      for (var i = 0; i < properties.length; i++) {
-        var name = properties[i];
-        if (name != "writable" && name != "enumerable" && name != "configurable") {
-          if (name in target) {
-            if (!(name in source) || typeof target[name] !== typeof source[name]) { 
-              throw new TypeException("Implementation of the property '" + property + "' is invalid");
-            } 
+      for (var propertyName in target) {
+        if (propertyName != "writable" && propertyName != "enumerable" 
+            && propertyName != "configurable") {
+          if (!(propertyName in source) || typeof target[propertyName] !== typeof source[propertyName]) {
+            throw new TypeException("Implementation of the property '" + propertyName + "' is invalid");
           }
         }
       }
+    },
+    writable: false,
+    enumerable: false,
+    configurable: false
+  },
+
+  /**
+   * @memberOf {class4js.Class}
+   * @static
+   * @private
+   * @method __getPropertyDescriptor
+   * @param {Object} object
+   * @param {String} propertyName
+   * @returns {Objecy}
+   */
+  __getPropertyDescriptor: {
+    value: function (object, propertyName) {
+      var descriptor = Object.getOwnPropertyDescriptor(object, propertyName); 
+      if (!descriptor) {
+        return Class.__getPropertyDescriptor(Object.getPrototypeOf(object), propertyName);
+      }
+      return descriptor;
     },
     writable: false,
     enumerable: false,
@@ -127,20 +146,23 @@ var Class = Object.create(null, {
   __instanceOf: {
     value: function (source, target) {
       if (source && target) {
-        for (var property in target) {
-          if (!(property in source)) {
-            throw new TypeException("Class doesn't implemet property: " + property);
+        for (var propertyName in target) {
+          if (!(propertyName in source)) {
+            throw new TypeException("Class doesn't implemet property: " + propertyName);
           } else {
-            var sourceDescriptor = Object.getOwnPropertyDescriptor(source, property);
-            var targetDescriptor = Object.getOwnPropertyDescriptor(target, property);
-            Class.__descriptorsAreEqual(property, sourceDescriptor, targetDescriptor);
+            var sourceDescriptor = Class.__getPropertyDescriptor(source, propertyName);
+            var targetDescriptor = Class.__getPropertyDescriptor(target, propertyName);
+            Class.__descriptorsAreEqual(propertyName, sourceDescriptor, targetDescriptor);
           }
         }
         return true;
       } else {
         throw new TypeException("Source or target is not set");
-      }   
-    } 
+      }
+    },
+    writable: false,
+    enumerable: false,
+    configurable: false
   },
 
   /**
@@ -221,8 +243,7 @@ var Class = Object.create(null, {
    * @method addExtension
    * @param {Function} extension
    */
-  addExtension: 
-  {
+  addExtension: {
     value: function (extension) {
       if (extension) {
         Class.__extensions.push(extension);